move parsers to amplify core
