+++ packages/amplify_core/lib/utils/index.dart
@@ -1,0 +1,1 @@
+export './parsers.dart';
