+++ packages/amplify_core/lib/utils/parsers.dart
@@ -24,3 +24,10 @@ String? enumToString(Object? o) =>
 // only to be used internally by amplify-flutter library
 T? enumFromString<T>(String? key, List<T> values) =>
     values.firstWhereOrNull((v) => key == enumToString(v));
+
+// only to be used internally by amplify-flutter library
+String upperSnakeCaseToCamelCase(String value) => value
+    .toLowerCase()
+    .split('_')
+    .map((e) => e[0].toUpperCase() + e.substring(1))
+    .join('');
